Add functions to remove unused map in Switch
Add functions to create Table/Group/Meter object on the Switch but not store
in the db (tableDb/groupDb/meterDb). This is because the application may
maintain its own cache for the Table/Group/Meter resources, and a storage in
ofnet is duplicated.

Add APIs in AppInterface to control if or not to initialize the flow graph and
the tlv map on the Switch.

Signed-off-by: wenyingd <wenyingd@vmware.com>